+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+
+## v0.1.1-dev (04 September 2024)
+- Bugs may appear. Please [open an issue](https://github.com/LeGeRyChEeSe/Sunshine-AIO/issues/new) if you have one.
+- Tool still under development.
+
+### Added
+- Shortcut to `option.txt` file to add custom resolutions/frame rates (See `7.` → `5.`)
+- Shortcut to Sunshine Settings. (See `7.` → `6.`)
+- Shortcut to Playnite. (See `7.` → `7.`)
+
+### Changed
+- Complete restructuration.
+- Code optimization. (Still optimization to be done)
+- Added a parameter to Sunshine Virtual Monitor command preps in Sunshine Settings for the Virtual Display Driver.
+
+### Fixed
+- Virtual Display Driver installation for Windows 10 and other issues. ([#4](https://github.com/LeGeRyChEeSe/Sunshine-AIO/issues/4))
+- Sunshine Virtual Monitor repository switched for my repo. It fixes the Windows 10 issue with Virtual Display Driver.
+
+
 ## v0.1.0-dev (28 August 2024)
 - First release of the Sunshine-AIO tool.
 - Bugs may appear. Please [open an issue](https://github.com/LeGeRyChEeSe/Sunshine-AIO/issues/new) if you have one.